pyproject.toml
[project]
name = "scala"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "typer>=0.21.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
